feat(magic): 新增构造解析、投影崩坏魔术及相关渲染器
- 新增“构造解析”魔术，支持通过准星或手持物品解析物品结构
- 新增“投影崩坏”魔术，可将投影出的物品作为爆炸性投掷物射出
- 为“投影崩坏”投掷物添加了动态膨胀的爆炸特效及专用渲染器
- 优化投影逻辑，支持副手投影并增加“双手已满”提示
- 调整魔术冷却时间从 1.0s 缩短至 0.5s
- 修正魔力值显示，将数值强制转换为整数以保持界面整洁
- 更新中英文语言文件以包含新增魔术的本地化文本

+++ src/main/java/net/xxxjk/TYPE_MOON_WORLD/procedures/Back_mana.java
@@ -8,6 +8,6 @@ public class Back_mana {
     public static String execute(Entity entity) {
         if (entity == null)
             return "";
-        return "当前魔力值：%s".formatted(entity.getData(TypeMoonWorldModVariables.PLAYER_VARIABLES).player_mana);
+        return "当前魔力值：%s".formatted((int)entity.getData(TypeMoonWorldModVariables.PLAYER_VARIABLES).player_mana);
     }
 }
