+++ src/main/java/net/xxxjk/TYPE_MOON_WORLD/procedures/Back_max_mana.java
@@ -8,6 +8,6 @@ public class Back_max_mana {
     public static String execute(Entity entity) {
         if (entity == null)
             return "";
-        return "魔力值上限：%s".formatted(entity.getData(TypeMoonWorldModVariables.PLAYER_VARIABLES).player_max_mana);
+        return "魔力值上限：%s".formatted((int)entity.getData(TypeMoonWorldModVariables.PLAYER_VARIABLES).player_max_mana);
     }
 }
